Personal_Projects/fraudley/gcp_frontend_deployment/requirements.txt
dill
gradio
joblib
numpy
pandas
lime
Pillow
requests
streamlit
xgboost==1.5.2
regex